+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.2.5 - 2018-09-11
 ### Fixed
-- Fixed a bug when creating import jobs
+- Fixed a bug when importing contacts
 
 ## 1.2.4 - 2018-09-11
 ### Fixed